portfolio-rag-backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
chromadb
python-dotenv